requirements.txt
networkx==3.2.1
numpy==1.26.2
psycopg2==2.9.9
python-dotenv==1.0.0
shapely==2.0.2
